requirements.txt
asgiref==3.8.1
Django==5.1.4
django-widget-tweaks==1.5.0
pillow==11.0.0
qrcode==8.0
sqlparse==0.5.3